+++ app/src/main/java/com/example/dell/petsforall/Home.java
@@ -1,15 +1,9 @@
 package com.example.dell.petsforall;
 
-import android.support.annotation.Nullable;
 import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import java.util.ArrayList;
-import java.util.List;
 
 public class Home extends AppCompatActivity {
     private TabLayout tabLayout;
@@ -21,10 +15,7 @@ public class Home extends AppCompatActivity {
         setContentView(R.layout.activity_home);
 
         tabLayout = (TabLayout) findViewById(R.id.tabHome);
-        tabLayout.setTabGravity(TabLayout.GRAVITY_FILL);
-
         viewPager = (ViewPager) findViewById(R.id.pagerHome);
-
         tabLayout.setupWithViewPager(viewPager);
         setupViewPager(viewPager);
 
@@ -32,39 +23,11 @@ public class Home extends AppCompatActivity {
 
     public void setupViewPager(ViewPager viewPager) {
         MyViewPagerAdapter myViewPagerAdapter = new MyViewPagerAdapter(getSupportFragmentManager());
+        myViewPagerAdapter.addFragmentPage(new DonateFragment(), "Doar");
+        myViewPagerAdapter.addFragmentPage(new AdoptFragment(), "Adotar");
         myViewPagerAdapter.addFragmentPage(new EditProfileFragment(), "Perfil");
 
         viewPager.setAdapter(myViewPagerAdapter);
-    }
-
-    public class MyViewPagerAdapter extends FragmentPagerAdapter {
-        private List<Fragment> fragments = new ArrayList<>();
-        private List<String> titles = new ArrayList<>();
-
-        public MyViewPagerAdapter(FragmentManager manager) {
-            super(manager);
-        }
-
-        public void addFragmentPage(Fragment fragment, String title) {
-            fragments.add(fragment);
-            titles.add(title);
-        }
-
-        @Override
-        public Fragment getItem(int position) {
-            return fragments.get(position);
-        }
-
-        @Nullable
-        @Override
-        public CharSequence getPageTitle(int position) {
-            return titles.get(position);
-        }
-
-        @Override
-        public int getCount() {
-            return titles.size();
-        }
     }
 
 }
