+++ app/src/main/res/layout/activity_home.xml
@@ -10,7 +10,11 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:id="@+id/tabHome"
-        android:background="@color/colorPrimary"
+        android:background="@color/colorPrimaryDark"
+        app:tabMode="fixed"
+        app:tabGravity="fill"
+        app:tabIndicatorColor="@android:color/holo_blue_bright"
+        app:tabIndicatorHeight="5dp"
         android:minHeight="?attr/actionBarSize"
         android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
