+++ app/src/main/res/layout/fragment_edit_profile.xml
@@ -60,14 +60,17 @@
 
         <Switch
             android:id="@+id/switch1"
-            android:layout_width="wrap_content"
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_alignParentStart="true"
             android:layout_alignParentTop="true"
+            android:layout_marginEnd="48dp"
             android:layout_marginLeft="16dp"
+            android:layout_marginRight="48dp"
             android:layout_marginStart="16dp"
             android:layout_marginTop="16dp"
             android:text="@string/label_cats"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="@+id/textView13"
             app:layout_constraintTop_toBottomOf="@+id/switch2" />
